Update integral length scale function (#376)
Discovered a problem with the original function while conducting data analysis of ADV data from Puget Sound.

Updating this function per the textbook definition of the integral time scale - it is listed as Equation 5.3 in this paper
(http://dx.doi.org/10.1098/rsta.2012.0196). The code I'm replacing defined the integral time scale as when the autocorrelation function decays to 1/e of its original value, which is less robust than finding the first zero-crossing of the autocorrelation function, as is standard.

The integral length scale [m] is simply the average flow speed [m/s] multiplied by the integral time scale [s].