Improve Nortek2 index file creator functions (#397)
Improves the nortek2 reader to handle block headers with a length of 12
bytes (which can happen for large data blocks according to the Nortek
integrator guide pdf - looking at the echo sounder here).

Improves/makes more robust the ability to search for a second global
configuration header if a file is corrupted/rewritten/appended to for
various edge cases.

Adds ability to read AD2CPs that are dual profiling with an echo
sounder, though the dual-profiling functionality is still effectively in
"beta".

Solution for Issue #391 and an update to PR #371 (which should have
solved Issue #392).