Read Sentinel V specific data packets (#396)
Co-authored-by: ululi1970

Issue #390 

Added a few Sentinel V specific code block readers. There are a few more
in the documentation that I did not add (that are not in this test
file), but I added "skip this many bytes" placeholders for if they're
needed in the future.

Also did a little cleanup and changed two ambiguous attribute names to
something more readable (hence a lot of test files changed), and I added
a hard stop to exit the forever loop problem. It may cause some partly
corrupted files that would otherwise be readable to fail, but we'll see.