Add WEC-Sim power performance example  (#395)
This PR adds an example that uses MHKiT and WEC-Sim together to do a
power performance assessment:
- utilize resource characterization in PacWave notebook
- write MCR cases for WEC-Sim
- load WEC performance
- assess WEC performance

Right now I represent the resource using the 32-cluster results in the
PacWave notebook and irregular waves in WEC-Sim. I was considering
writing all 2000+ wave cases in that example to a WEC-Sim MCR file so
that we can demonstrate
`mhkit.wave.performance.power_performance_workflow()` but decided it
would be too expensive, even if regular waves were used in WEC-Sim.

I also have a half-finished example that shows how to write a
directional spectra (like in `directional_waves.ipynb`) for WEC-Sim. But
the WEC-Sim 'full directional spectra" feature needs documentation and
an example first.

The PR also includes a function to make the xarray representation of
WEC-Sim data a lot more useful

---------

Co-authored-by: Mohamed Shabara <84589678+MShabara@users.noreply.github.com>