Short list of VMDAS updates (#405)
Discovered from a recent vessel survey that I misunderstood TDI's
documentation on a couple of VMDAS parameters. The speed and direction
parameters I thought it was pulling directly from the GPS it is
calculating from the change in latitude/longitude over time. I renamed
them and added the two variables that are actually directly from the
GPS.

Also discovered that VMDAS cannot take an RMC NMEA sentence, so the GPS
timestamps are lacking a datestamp. Which is not helpful if your ADCP's
clock is on local time and the GPS is always in UTC. So I added a code
block that checks if the GPS UTC time rolled over from 23:59:59 to
00:00:00 and manually updated the day.

---------

Co-authored-by: ssolson <sterlingolson@yahoo.com>
Co-authored-by: akeeste <akeeste@sandia.gov>