Allow user to specify universal Kolmogorov constant for TKE dissipation rate function (#406)
It was pointed out to me that a universal constant of 0.5 was hardcoded
into dolfyn when calculating dissipation rate using the Lumley and
Terray method. According to turbulence theory and historical
experimental validation, this is the value to use for the streamwise
velocity spectra. According to the same, the constant is different for
velocity spectra that are orthogonal to the streamwise direction -
usually taken to be somewhere between 0.65 and 0.69.

This PR allows the user to set this constant via function input.

Pope, Stephen B. "Turbulent flows." Measurement Science and Technology
12.11 (2001): 2020-2021. pp 231-232.