Nortek Dual Profile Dataset Rotation (#414)
Discovered this problem working on PacWave data, where the second
dataset, which contains instrument-averaged velocity measurements,
errors out when trying to use dolfyn's 'rotate' functionality.

This PR ensures the second profile from Nortek instrument can be rotated
like that of the first. This is to fix a somewhat hacky fix from several
months back, and doesn't require the user to make additional changes to
the 2nd dataset beyond that what dolfyn already does.