Fix pip module tests (#423)
Status: Ready

This fixes test failures related to module loading using `pip install
"mhkit[<module>]"` syntax.

There are some modules that have interdependencies with other modules,
i.e., loads module needs wave module code that are causing test
failures.

The root cause of test failures is specific packages that are not
specified in pyproject.toml for a specific module, but are called when
the tests are run.

The fix is to add lazy loading for the `wave.io` module and reconfigure
the modules in pyproject.toml to capture the a more correct dependency
list derived from running the tests.

Also the command that runs these tests was changed from `pip install -e
"mhkit[<module>]" .` to `pip install -e ".[<module>]"` because the tests
fail with `-e "mhkit"` but succeed with `-e "."`. I suspect that this
changes the way mhkit is installed, but I don't know what the specific
differences are.

---------

Co-authored-by: jmcvey3 <53623232+jmcvey3@users.noreply.github.com>
Co-authored-by: akeeste <akeeste@sandia.gov>